feat(gui): honour the census tier for the Shortcuts advanced half
Owner ruling: where the census and the renderer disagreed about a
tier, the census wins. Four families said `.showMore` — the layer
strip, its menu-bar icon row, the switch-layer rows and "Show
shortcuts panel" — while the renderer drew all four at rest, and
set-equality could not see it because membership was right.

So they move behind disclosures, which is what `.showMore` means:
one interaction away. The strip, its icon row and the switch rows
become one **Layers** card, titled for the container rather than
for one of its parts, sitting with the other advanced cards; the
General keys row gets its own drawer. The ⌃⌥K panel keeps
`shortcuts.section.switch_layers` for its band, which really is
only the switch rows.

The strip is also the editor's layer SELECTOR, so hiding it hides
what every group below binds into. The header keeps the fact and
gives up the control: a quiet "Editing the "x" layer" label,
shown only once a second layer exists — which is the shape the
design already specified for the panel's layer chip, a label and
not a second way to change the selection.

`LayerStripEditor` becomes its own view because it owns four
pieces of `@State`, and `@State` on a struct never mounted in the
hierarchy silently does nothing. That also takes `ShortcutsHeader`
from 327 lines to 118 on a real seam — the earlier attempt to
split it for size alone was reverted, because it would have meant
widening `private` state to satisfy a soft guideline.
