+++ Sources/KiwiDeskCore/Resources/Locales/en.json
@@ -667,6 +667,7 @@
   "shortcuts.base_layer_protected": "Base layers can't be removed here",
   "shortcuts.choose_app": "Choose app…",
   "shortcuts.delete_layer": "Delete layer",
+  "shortcuts.editing_layer": "Editing the “%1$@” layer",
   "shortcuts.go_to_space": "Go to space",
   "shortcuts.import": "Import from init.lua…",
   "shortcuts.import.help": "Reads the shortcuts active in init.lua and adds them here, matching each combo. Known actions sort into the groups below; anything else lands in Advanced. Review, then Save.",
@@ -699,6 +700,7 @@
   "shortcuts.section.focus": "Focus",
   "shortcuts.section.general": "General",
   "shortcuts.section.inactive": "Inactive shortcuts",
+  "shortcuts.section.layers": "Layers",
   "shortcuts.section.move_windows": "Move windows",
   "shortcuts.section.open_applications": "Open applications",
   "shortcuts.section.size_float": "Size & float",
